requirements.txt
python-telegram-bot==22.4
yt-dlp
aiofiles
